add Events

+++ service/Service.js
@@ -6,32 +6,103 @@ class Service {
     
     constructor (db) {
         this.db = db;
+
+        this.beforeCreateHandler = null;
+        this.afterCreateHandler = null;
+
+        this.beforeUpdateHandler = null;
+        this.afterUpdateHandler = null;
+
+        this.beforeDeleteHandler = null;
+        this.afterDeleteHandler = null;
+    }
+
+    set onBeforeCreate(handler) {
+        this.beforeCreateHandler = handler;
+    }
+
+    set onAfterCreate(handler) {
+        this.afterCreateHandler = handler;
+    }
+
+    set onBeforeUpdate(handler) {
+        this.beforeUpdateHandler = handler;
+    }
+
+    set onAfterUpdate(handler) {
+        this.afterUpdateHandler = handler;
+    }
+
+    set onBeforeDelete(handler) {
+        this.beforeDeleteHandler = handler;
+    }
+
+    set onAfterDelete(handler) {
+        this.afterDeleteHandler = handler;
     }
 
     findAll() {
-        try {
             return this.db.find();
-        } 
-        catch(e) {
-            console.log(e);
-        }
-
     }
     
     findById(id) {
         return this.db.findById(id);
     }
 
+
     create(data) {
-        return this.db.create(data);
+
+        let proceed = false;
+
+        if (this.beforeCreateHandler !== null)
+            proceed = this.beforeCreateHandler(data);
+     
+        if (!proceed) 
+            return null;
+
+        let result = this.db.create(data);
+
+        if (this.afterCreateHandler !== null)
+            this.afterCreateHandler(data, result);
+
+        return result;
     }
     
     update(data) {
-        return this.db.findByIdAndUpdate(this.data.id, data);
+
+        let proceed = false;
+
+        if (this.beforeUpdateHandler !== null)
+            proceed = this.beforeUpdateHandler(data);
+
+        if (!proceed) 
+            return null;
+
+        let result = this.db.findByIdAndUpdate(this.data.id, data);
+
+        if (this.afterUpdateHandler !== null)
+            this.afterUpdateHandler(data, result);
+
+        return result;
+
     }
     
     delete(id) {
-        return this.db.findByIdAndDelete(id);
+
+        let proceed = false;
+
+        if (this.beforeDeleteHandler !== null)
+            proceed = this.beforeDeleteHandler(id);
+
+        if (!proceed) 
+            return null;
+    
+        let result = this.db.findByIdAndDelete(id);
+
+        if (this.afterDeleteHandler !== null)
+            this.afterDeleteHandler(id, result);
+
+        return result;
     }
 
 }
